修改了resize－cluster脚本，可部分自动配置信息

+++ resize-cluster.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
 # N is the node number of hadoop cluster
-N=$1
 
 if [ $# = 0 ]
 then
@@ -10,19 +9,35 @@ then
 fi
 
 # change slaves file
-i=1
-rm config/slaves
-while [ $i -lt $N ]
-do
-	echo "hadoop-slave$i" >> config/slaves
-	((i++))
-done 
+changeNodeNums() {
+    Num=$1
+    echo "Chang Hadoop Slave numberbers !"
+    if [ -e config/slaves ];then
+        rm config/slaves
+    fi
+    i=1
+    while [ $i -le $Num ]
+    do
+        echo "hadoop-slave${i}" >> config/slaves
+        i=$(($i+1))
+    done
+}
 
-echo ""
+removeImage() {
+    echo "Before create, we should clear the exists hadoop !"
+    sudo docker images | grep "hadoop" | awk -F' ' '{ print $3 }'
+    if [ $? -ne 0 ];then
+        return 1
+    else
+        imagesid=`sudo docker images | grep "hadoop" | awk -F' ' '{ print $3 }'`
+        sudo docker rmi -f ${imagesid}
+        return 0
+    fi
+}
 
+N=$1
 echo -e "\nbuild docker hadoop image\n"
-
-# rebuild kiwenlau/hadoop image
-sudo docker build -t kiwenlau/hadoop:1.0 .
-
-echo ""
+changeNodeNums ${N}
+removeImage
+sudo docker build -t "dockerfile/ubuntu14.04:hadoop" .
+sudo docker images
